project 4 refactor, add ImageUtils
