+++ src/main/java/edu/project4/transformations/DiskTransformation.java
@@ -1,4 +1,4 @@
-package edu.project4.Transformations;
+package edu.project4.transformations;
 
 import edu.project4.Point;
 import static java.lang.Math.PI;
